cleanup script now crucial in case previous run errored

+++ lib/blm_cleanup.py
@@ -47,7 +47,8 @@ def main(ipath):
     # Clean up files
     # --------------------------------------------------------------------------------
     os.remove(os.path.join(OutDir, 'nb.txt'))
-    shutil.rmtree(os.path.join(OutDir, 'tmp'))
+    if os.path.isdir(OutDir):
+        shutil.rmtree(os.path.join(OutDir, 'tmp'))
 
 
     print('Analysis complete!')
